This commit tried to get rid of explicit odemis package references in import statements, but that is not possible at this time.
The 'model' module gives a conflict when launching the XRCed editor, so an
implicit 'odemis.model' import is needed to prevent naming conflicts.

As a side effect, the PYTHONPATH for the GUI needs to include both the
directory containing the main 'odemis' source directory as well as
the directory itself.