model: return NotApplicableError if choices or range is not existing
the property used to return AttributeError, as expected, but Python would
get look in the Proxy class and return a RemoteMethod, instead of an exception.
To force the exception, we need to raise a different exception type.

Note: the alternative would be to extend __PyroAttribute of the proxy, but I don't
like to mess with frozensets