model: don't terminate() component on __del__
the main problem is that the environment where __del__ is executed is ensure, because many parts of the program
might be already gone. In particular, logging might fail with an exception.

In addition, __del__ prevents garbage collections, which might sometimes be useful.

As component usage _must_ be finished by a call to terminate, that shouldn't be necessary anyway.