semcomedi: refactor
put separate parts in separte methods to clean up the code and make it slightly understandable

it's not very clear how the convertion from raw data to physical value will be, it would be best
that it takes place while reading (we have 800ns per read). Maybe memmap will make it easier. it
should also avoid the margin data.