This commit prevents the custom Slider controls from firing an 'unlimited' amount of events in a short time. They will now fire once every 50 ms at the most.
This limitation also seems to have resolved a bug which was causing an
infinite loop with Sliders that were connected to VigilantAttributes. This
was not caused by float type inaccuracies, which were encountered earlier
and have been resolved. What the actual cause was, is not fully clear, but
the control would start going back and forth between two values, e.g.
0.157049655914 and 0.151920482516.