gui: unsubscribe/subscribe to dataflow when the dwell time is changed
Changing the dwell time of the HardwareComponent doesn't cancel the current
acquisition. If the dwell time is large, it will cause a long acquisition which
is running even if it was a mistake by the user.

Reduce the user's surprise by automatically cancelling acquisitions when the
dwell time is big.