pvcam: fix unresponsive camera after abort
seems like camera becomes unresponsive if the abort happens at the end of the acquisition (=during readout?)

There is two ways to avoid this: wait for this (small) period, or close/open the camera again.
Let's wait for the acquisition to end, whenever it seems it should be soon. That mainly avoids
the very long close/open cycle on ST133.

Also fix the test cases with binning being a tuple