semcomedi: bug fixes
fix a bug when sub/unsub fast. We need to ensure that the cancel of the reader and writer is done after the preparation, and before the set_resting_position

workaround a bug when writting one pixel at a time, the driver seems to keep running. Still needs more investigation because I think it must have been working before.