gui: refactor toolbar
simplify toolbar class to just have one implementation, which supports any tool.

It has only one public method: AddTool()
The tab controller is in charge of calling the class with the right tools at init.

Also link the "fit zoom to content" functionality

still needs a bit of update on the XRC file

It would be even better if we could get rid completely of the XRC based children of the panel,
then the code could be simplified further.