semcomedi: fix mmapreader to work with NI driver
found out finally that mmap needs comedi.poll() to work correctl with the NI driver
probably a bug in the NI driver, but that works around it nicely.

Then we have a workaround the kernel crash with cancel() and read()
Still currently the normal reader is the default. Will need to switch from
Reader() to MMapReader() if the fixes don't come to Ubuntu kernel soon.