make the GUI open correctly fluorescence OME-TIFF
MD_OUT_WL needs to be recorded. For now, the only way to do that easily with multi-band filter is in the
GUI, which know which of the band we actually use.

Special static streams for fluo

Detect data with metadata MD_IN_WL and MD_OUT_WL as fluoresence and use StaticFluoStreams

Handle correctly readonly wavelengths in the stream entries