gui: small improvement with colour management
Improve change_brightness() to support better dark or light colour: "blend" instead of adding.

remove dependency on StepColour(): all the usages can be move to change_brightness

add borders to the visualSliderRange window

don't change the colour of the visualSliderRange when disabled (but the borders of the window) -> easier to read

also improve SetRange of visualSliderRange to detect bad values of min/max and put less warnings.