gui: fix/improve a bunch of size and mpp related problems in the canvas
DblMicCanvas listens to view_pos, so that it's possible to change the position just by updating the VA (such as in the acquisition dialogue).
Don't move if NoDragNoFocus set => was a thinko: or/and
Automatically adapt the view to the new image when opening a new acquisition file.
remove some non needed mpwu complexity: fix it to 1. In the future, it should be possible to completely get read of it.