gui: fix overlay selection with Y inverted
Ensure after every conversion that the top/bottom and left/right are correctly ordered,
so that if Y is in opposite direction, it doesn't aftect the code. It also fixes some
corner cases when the user was dragging to redimension a selection from one side to the
other side.

Also clean up code in RepetitionSelectionOverlay by keeping functions small.