gui: ensure the "acquire image" is always enabled when it's possible
it would get disabled when opening an image:
 * open image
 * clear the streams of the panel
 * send message no streams
 * disable button
 * change back to stream tab
=> button is still disabled

the current fix is to also pass the tab for which the stream is about. So we don't disable the button when it's about other tabs

Probably the best way, for long term is to remove all the pubsub infrastructure, and only use VAs. StreamTree could be VA like, with events whenever it changes.