model: ensure that ListVA also works when used remotely
Need to convert back the special NotifyingList into a normal list, to allow pickling.
New Proxy that converts the normal list into a special notifying list every time it receives the value.

It's propably dog slow. But we care more about correct behaviour than speed, and normally the lists we use are < 10 elements, so it should be fairly cheap anyway.