backend: add support for configuring the pole position in the SPARC
This position cannot be computed unless the position of the lens (X/Y) is known.
After discussion with Andries & Wouter, it seems actually easier for the user to
just enter the position of the pole on the CCD (when the lens is set).

So extend the Lens component to support bearing this information, and mdupdater
to set the metadata of the image.