gui: inverse the Y axis of the SPARC AR
The convention is to have the Y axis to the top, which means in manual
alignment mode, which has no lens, it should look toward down. IOW, we want
the bar of the D on the top in normal acquisition, so it should be down on the
alignment.

Still needs to be tested on the real hardware to check everything has been correctly
inverted.