semcomedi: fix race condition when starting acquisition immediately after stopping it
The code would test wrongly for the thread to still be active and give up if it is.
However, if the acquisition has just finished, in some case the thread might still be up.
It will die very soon.
This is handled by _wait_acquisition_stopped() anyway, so trust it.