driver: add .softwareTrigger event
Special event that is now required on all components which have a synchronisable dataflow.

Although not purely required, this avoids the frontends to have to provide an Event whenever they need
to do software trigger.
Typically useful for when wanting to do multiple acquisitions at software-decided moments.

Will be used in SEM/CCD acquisition