andorcam2: add info from Andor about EMCCD gain
"I've tested this with an Ultra here and have reproduced your issue.  Real gain should be selectable using SetEMGainMode(3), this is the case for iXon3 which has the same sensor.  So I have raised this issue with the Ultra/SDK team. Linear gain is very similar to RealGain.  The main difference is that it does not compensate for EM Gain aging. For now this should give you similar results until I get some more information from the Ultra/SDK team."