gui model: change tab switching to use a VA
Instead of having a special singleton (tab controller) and a fake list, use a VA,
more or less similarly to .focussedView.

We even go further with .tab to also contain all the available tabs on .tab.
Probably we could do the same on .focussedView/.views.

The main problem of this new solution is that the model is a bit "dirtied" by referencing a Tab,
which is a not a model object.

Directly use the new VA to fix a bug in snapshot: acquisition is now of the _current_ tab.