dataio: handle correctly non-ascii filenames
The only way in Python to support every filnames is to use unicode strings.
HDF5 was doing fine fine, but TIFF needs to explicitly go back to the current filesystem encoding.

Now expects the filename to always be unicode, to force the correct handling.
Update the extensions to be unicode.

Add also add test cases.