gui: handle better non-ascii filenames
Filenames should be always encoded in unicode to work correctly.

So force all the filename-related functions (that I could find) to use unicode.

That fixes a bug when the user would try to save a file in a directory with non-ascii characters.

Also detect if the currect directory doesn't exists (anymore), and fallback to the default directory.

This fixes another bug which could happen if the directory was moved/renamed.