driver andorshrk: add simulator + move AndorSpec
Very simple simulator that allows basic testing of the class.

Move AndorSpec to andorshrk, so that andorcam2 doesn't need to know anything about spectrometers.
Also modify it so that it get a second child corresponding to the AndorCam2

Modify AndorCam2 to follow the same style as Shamrock for the simulator: device = "fake" creates a simulated device.