gui: don't turn on the light with maximum power
it can be pretty bad to have the full power on the sample.
Some users suggested to just leave the power to 0, but I'm not fully convinced as it'd mean that turning on the
optical microscope doesn't work, and the power still needs to be on. So as a middle-ground I'm using 10% of full power.