gui: reorganise file calibration change
now there is almost a direct connector between the file VA and the file widget (handled in the settings controller).

In addition, the settings controller provide the possibility to add a "setter" whenever the user changes the file.
This way, the tab controller can first check that the calibration is correct and refuse if necessary.

We also get/set the configuration file automatically from the VA. This could be implemented to a couple more settings if
that turns out to work well.

All this simplifies a bit the code but also ensure that the calibration file displayed is the correct one when an
error happens in the loading. It also adds tool tip for the label.