gui sparc: add count over time graph for spectrometer
As requested from user feedback in order to allow alignment of the optic fiber to the spectrometer.

For now it BREAKS the only SPARC system we have: the Andor SR303i doesn't allow changing the center wavelength
while acquiring spectrums. As now the spectrometer is constantly acquiring, it's impossible to change the
center wavelength.