andorcam2 + andorshrk: share the hardware access
pretty ugly, as it's adding quite a lot of complexity to the already complex acquisition thread
In addition, it's hard to track when the hardware is actually acquiring, and to ensure we only acquire/release once the lock.
So this is actually not supported when the acquisition is synchronised.

Anyway, that should be enough to deal with the one annoying case: during acquisition setup on the SPARC.