acq find_overlay: change electron coordinate computation
this thing is absolutely confusing. It's not clear whether the referential should
be from the borders of the SEM FoV or the actual centers of the spots.

Testing on a static optical image of 7x7, it seems it should be the actual
centers, so that's what we now use. However it seemed to be working
previously with real images... we need more testing!

Also simplify a bit the code.