driver semcomedi: fix placement of pixel centers
The e-beam position of each pixel must be at the center.
The computation of the coordinate was wrong, leading to the ebeam being placed almost
at the border of the pixel. This was affecting very large scales, such as when doing fine alignment.

Note that spot mode was not affected.