gui/acq stream: ensure the range of intensityRange is synchronised with the data
the difficulty is that histogram is computed delayed, but can be used to compute the values.

So now force histogram computation if the range changes.

Also ensure that the GUI doesn't update the range too often in the controls.
Especially it was preventing the user from ever being able to modify the values.
Prevent crashes of the GUI by putting the changes in a call_after block.