gui canvas: avoid blinking empty rendering
the drawing routine runs in a separate thread. So it could happen that while updating
the list of images, it would pick the list on the way to be constructed.
Instead, ensure that the new list is set only once fully constructed, and then ensure
we only use this list in the drawing method.