config: clean up config files
remove some non useful config files
move the config files specific to some sites to the examples/ directory.
This should make it clear they should not be used as-is, but first copied.
The main problem of using them as-is is that the odemis debian package will update them
silently on the next version.