driver andorcam3: add translation VA
the newer SDK and hardware supports ROI anywhere on the sensor.
Let's allow the user to change the center of the ROI with the .translation VA.

Works mainly the same as for the SEM, expected it must be ints.

Also add support for the transpose options in DigitalCamera, and a test case.

Untested on real hardware.... but should probably not have any effect as long as
it's not changed at least.

Will be used by the bleaching script.