hdf5: handle RGB data better
If data is explicitly RGB (CYX or YXC), then save as CYX, and read it back correctly.

That allows to actually be able to see the image in hdfview, and detect when reading back that
it is a RGB data.
It doesn't work in Huygens... but almost could.

Also restrict further the grouping of data to only group over channel.
That's only case where that matters, and it avoids trying to be too clever with data given.