gui model/tab: reorganise the instrument state to be part of the tab model
The new idea is that just one tab is in charge of handling the _live_ microscope state.
So we can move the states to the tab data.
This allows to easily have access to the streams, from the chamber functions.

We might be able to simplify further by getting rid of the states completely and moving the code
to the tab controller itself.