gui: move chamber management to microscope controller
now we can do things more properly: the control part is in the controller.
In particular, this allows to move most code together, so ensure the order of
execution.

It also fixes some problems with the overview acquisition when starting in this mode.

If no chamber at all, then just highlight lens alignment tab at init.