driver semcomedi: handle dwell time > 15 s
Dwell time used to be limited by the time represented in ns on 32 bits.
After fixing this limit, it was limited by the number of samples
represented on 24 bits. On the NI 6521, it was effectively limiting the
dwell time to 15s... which some people hitted.

Break this limit by cutting each pixel into sub-acquisitions.