acq mng + gui: change acq manager future to return result + exception
We return the result/error in a very unusual way from acq.acquire()'s future:
result + exception

The reason is that sometimes the acquisition fails half-way.
All the data already acquired might still be useful (and in some cases, it
might be impossible to acquire again due to sample degradation).

So this allows to report errors while still returning the data acquired so far.

Also update the GUI to take advantage of this (iow, save the data acquired in
case of exception).