driver tmcm: improve referencing and fix basic simulated version
Upload the programs once at init, and never again

Change the procedure to do the fast referencing with negative last,
so the slow procedure is normally only ran in one direction.

Fix error with setting interrupt for axis != 0

Fix future for referencing

Add test cases for referencing

Also extend a little bit the simulator so that it's possible to run it as is.