model: add support for multiple parents for component created by delegation
Introduce the "creator" attribute: specifies the component that will be used for creation.

If a component has only one parent, it can be implicitly filled, so the current working microscope files are still valid.

Also add support for the CombinedActuator to components without .speed, to keep the test microscope file minimal.