config: do not start odemis with a terminal by default
Try to make it less worrying/ugly for the users.

If only the GUI starts, don't display anything.

When starting the backend, just use popup notification for normal cases,
and display the log only if the backend failed to start.
Still use a terminal if it's LOGLEVEL=2 and backend starts.