driver tmcm: detect if the device has no power for the motors
The device look fine, but the motors will not move, which is kind of important.
The only way I found to detect is to rely on the fact the programs don't run if
the device is not powered.

Also timeout from the driver side in case the reference program never ends,
which could happen for instance if the power suddenly goes down during referencing.