gui: wait for focus move to be over before sending new move request
If the focus move is slower than the time between two move requests (used to be 10 Hz),
the move requests will accumulate, and cause a delay between what the user asks and the
actual hardware move.

I've tried making the drivers more clever, by accumulating requests, but then we end up
with lots of complex code for handling cancellation of multiple futures simultaneously.

So instead we change the GUI code to explicitly wait for the move to be over before
requesting more. This allows to automatically adapt to the speed of the hardware.

Also simplify the code as we _never_ have multiple focus (for now), and if we ever do
it will not be horizontal vs vertical.