acq stream: modify FluoStream .excitation and .emission to reflect hardware capabilities
In order to be able to show in the GUI the hardware settings, we need to express
the emission and excitation directly as the hardware settings.

This avoids the stream to try to be too clever, so that the GUI will be using the util.fluo
functions to try to be clever _and_ warn the user if the hardware cannot do something.