driver andorcam3: don't wait too long between frames
as the camera automatically starts the next frame acquisition, the time to wait for the next frame
can be less than exposure time + readout time. So waiting that long was constantly adding
a delay, that kept building up until the SDK queue was full. Instead, we just wait for the next
frame immediatly.

also add setup to read metadata (=timestamp), which is what allowed to detect the problem