driver andorcam3: improve acquisition speed and ensure settings are fixed
Small tweakings to improve acquisition speed.
It will still fill up the hardware queue if the frame rate is too high.
So far the tentatives to handle full hardware queue in a simple way failed.

Fix shutter to global, as it ensures all the pixels show the same thing at the same time.

Also explicitly set the framerate to the maximum. It seems that with the Zyla USB, sometimes
it remembers the old framerate, which is counter intuitive for the user.