gui config: make calibration config work, and other improvements
Finish the AcquisitionConfig class, with all the values saved and restored correctly

Simplify a bit the code.

Now .set() always implies .write(). It's "suboptimal" when changing multiple options
at the same time but this happens rarely and the config files are so small that it costs
not much anyway.

Last but not least, create test cases.