gui canvas: hack to make sure the whole SEM image is displayed
return as HFW the smallest dimension of the FoV.

That works because the SEM image usually square. This hacks needs to
be replaced later by something cleaner. Moving the cleverness in the
controller would allow to make it depend on the SEM image ratio...
but it's for later as it involves quite a lot more changes.