driver tmcm: if the checksum is incorrect, just assume the answer says it's all fine
From time to time, the message received is garbled.
If we skip it, eventually the whole instruction is considered failed
and an exception is raised.

As it seems it's just randomly garbled, and not caused by an actual
error in the controller, now we consider such garbled answer to be
exactly the message we expected.