gui microscope: simplify selection of chamber button faces
Don't rely on the role, but just on the properties of the microscope
to select the button faces of the chamber button.

So far it works better, as it detects the Phenom now the same way as
the Delphi. Once we support another douzain of SEM, we might need to
review this again ;-)