util stage: Generalise and simplify Antibacklash stage
Support any number of axes, and don't ask for their names, as they
are the same as the axes of the child.

Don't move axes which are not involved in the original call.

Don't create our own position VA, but use directly the one of the child.

Introduce test cases.