driver stage: Simplify CoupledStage behaviour and code
The master and slave stages now go to the _same_ position (instead of
opposite position). Note that this means 2 things need to be changed
for the Delphi: the microscope files now have the optical stage inverted,
(to keep the axes going opposite direction), and the translation calibration
values must be inverted too.

After updating the metadata, the stages are immediately synchronised to
the same position (instead of waiting for the next move).

Also simplify some code, and update the test cases.