driver tmcm: read temperature from analogue inputs
Apparently, the Delphi has two temperature sensors connected to the two
analogue inputs.

So read the temperature and provide it via a VA.

It's disabled by default as it only works if an sensor is present, and it's
in the correct format.

Still needs to be tested on a real hardware.

Also, we need to decide if both sensors are worthy to report.