gui settings: fix radio button not linked to the right value
If choices is a set (or a dict), it will not have any order, so we cannot rely on it.

That's why choices_fmt contains both original values, and formatted value. => use it.

Also fix missing unit prefix.

All these were noticable on the readout rate of the fake andorcam2.