odemisd: fix log file rotating multiple times in a row
due to multiprocesses, the RotatingFileHandler would be duplicated, and
when a log reached the limit, it would be multiple times rotated.

So instead, keep things simple, and rotate manually, just at the beginning.

The more fancy way would be to set the loggers to a pipe, and have this pipe
saved into a file (and rotated) in a separate process. For now, simple should be
enough.