gui viewport: make magnification easier to understand
Suggestion from Lennard.

The magnification is always the same (magical) number, indepedent of the number of
images displayed. If only one MPP, then also indicate the digital zoom.

That is more involved than it sounds because the text got bigger, so we need some
cleverness to automatically resize the textctrl.