gui: overview fixes and clean up
Add the overview also to optical-only and em-only microscopes.

Fix the stage not moving when clicking on the overview when no chamber is present
 => we now activate the overlay also when the chamber stage is unknown.

Don't show the cross-hair cursor if it's not possible to move (by actually
disabling the overlay when it shouldn't be possible to move).

Move the clearing of the history list VA (when the chamber pumps up) to the
microscope controller

Let the history overlay decide when it should force a redraw (=when the VA updates)

Move the code remove the crosshair by default to the tab controller (instead of viewport).

Remove unused code.