gui canvas: fix cursor resetting when changing focus
PointsOverlay would always reset the cursor, even when it's not active and
not in used. That caused the focus cursor to be reset as soon as the mouse moved.

=> Fix PointsOverlay

Also don't add PointsOverlay at all to views which have no chance to ever use it
(based on the choices of the .tool VA)

And fix the reset of the cursor to any mouse up (not just left up)