gui viewport: fix spectrum plot not updating
It used to rely on should_update VA to detect a new stream is present.
This VA was not made for that, and since we don't autoplay (static) streams anymore,
it got broken.

Instead, fix it by using some slighly closer to the meaning (but still not really correct),
which is to listen to view.lastUpdate, which is updated once everytime a new static stream
is added.