gui stream controller: force the current view to contain the stream when playing it
When starting a stream, if it's not currently visible, it's weird to be sent to another view,
which has the stream visible.

It's much simpler for the  mind if the stream automatically becomes visible in such a case.

Of course, the view has to be compatible. So that if starting the SEM while watching an optical
view, the user is still sent to the SEM view.