gui microscope: reorganise sample loading on Delphi
The main problem with the previous organisation was that actually the Phenom would never load a
non-registered sample holder. The whole procedure would flat-out stop there.

Now initiate the registration procedure much earlier so that might work.

Also rearrange some methods to be able to share code with the parent function.