model: extend container creating timeout to 120s
In some cases (like the Andor cameras), the initialisation can be really long. Sometimes it goes over 60s and then init fails.
Luckily, the camera is faster to initialise on the next times so Odemis usually succeeds in starting the second time.

It's still better to not timeout on such cases.