backend: clean up
make the CONTAINER_DISABLE code much more similar to the standard code.

Use exceptions to pass errors.

Move the management of the .Root object to the container, so that the object
and the getter and setter are declared in the same class (instead of only
being present in one child).

change signature of createInNewContainer to return both the container and the component.
All the callers had to look for the container afterwards, so better just pass it.