gui: Separate MicroscopeView from OverviewView
now the logic of moving the view to where is the stage can be at the right place, in the view.

Also allows to really never move the view position of the overview unless actually needed (which will
be the case once we support the recentering based on the old positions)

Also simplifies the OverviewViewport to not hunt for a view with a stage.

Need to introduce "cls" argument in the view controller to select which kind of view needs to be created.