driver pigcs: use MM as unit and don't override num/denom
the num/denom is actually a configuration value that matters, so we can't just override it.

Now expect that the unit is MM and bail out if it's not.

It should not cause problem with the current C867-CL we have as they are already configured in MM.

Also make sure to leave the initial speed as is.