driver pigcs: try to turn off encoder when not used
UNTESTED code that might break C867 CL !

Having the encoders always turn on can be annoying because they heat up the stage.

So try to turn them off when not in used (= not moving).

This is untested code for turning on/off the encoders, and untested code for ensuring the
encoders are really off iff the stage is supposed to not move.

Also do some small clean ups and add some comments about ONT failing after stop.