driver comedi_simple: correctly handle new comedi module 0.10.2
The new python module comedi already has the function names without the comedi_*.
So we don't need to handle this anymore, just the error codes.
With such version, calling the comedi* functions will not work.

Also improve exception message.
Also improve the wrappers to not depend on partial (make the calls slightly faster).

It should also be compatible with the older module versions.