driver pigcs: various small bug fixes (for CL moves)
After trials with the hardware, it seems it's not needed to turn on the encoder when just reading the position.

Update various comments. In particular, we need to handle time estimation better.

Don't fail the ActionManager if the action raises an exception (=> it should eventually be set as a future's exception)

Fix call to SendOrderCommand() in case of major connection failure.

Increase timeout over IP connection for avoiding timeouts when using WiFi connection.