gui lens alignement: use the right axes for the aligner
Detect if X/Y axes are there (SECOMv2) and directly link the aligner to the image instead
of using a ConvertStage.

For the SECOMv1, the behaviour should be unchanged.
However, it's all a bit fishy because it seems that the values correspond to the
inverse of the convention (A/B values are opposite). So some test should be done on
real hardware.

Also needs to be tested on the SECOMv2.