gui log: save log into log file from the very beginning
The log file would start at the same time as the standard output is switch to the log panel.
In case of error at initialisation it's really too late. So start the log file as soon as the log is initialised.

Also fix a exception in the exception handler when ending the program.