gui views controller: fix thumbnails updating from wrong viewport in analysis tab
As the connection btn <-> viewport can now change, the unsubcription from the previous viewport
must be carefully done.
Without it, after opening multiple acquisitions of different types, the thumbnails of the btns would
be updated by several viewports.

So introduce a new _unsubscribe() function that allows to do the right thing at the right time.