gui comp: fix merge slider icon order
The order the are drawn does affect the meaning of the merge ratio.
So make sure that the order set by the MicroscopeCanvas agrees with
the expected order by the viewport.

The right way would be to code everything in the StreamTree.
The current hack is to hard code the only 3 possibilities:
 - Optical/SEM
 - Spectrum/SEM
 - SEM/SEM