gui config: fix sorting of VAs and scale creation
The sorting of the VAs was based on complex code, which in some cases, could lead to:
  File "/home/secom/development/odemis/src/odemis/gui/cont/settings.py", line 808, in add_hw_component
    vigil_attrs[i], vigil_attrs[j] = vigil_attrs[j], vigil_attrs[i]
IndexError: list index out of range

Now rewrite it based on sort()

Also fix the controls using 1d binning and 1std binning which had both the new choices
function and the old type. This prevented to actually work at all.
=> drop the old types.