gui stream: fix fluo stream tint reset on acquire image
When creating a new stream entry for a fluo stream, if the tint was not
the default one and no dye was selected, then it the initialisation of
the widget would reset the tint to the default colour.

Modify the way, so that the tint is updated only when explicitly changing
the widgets.