util img: fix DataArray2RGB not working on uint8
What sounded the easiest turned out to cause a lot of problems:
bytescale just return the original data if it's (already) uint8.

In addition, numpy.clip() uses the dtype of the range (not the data) as output.

This second problem hid the first one when irange was made of floats (which is the
case in the GUI if the auto BC is used. So this led to the stange behaviour of
the function working only if auto BC was in use.

So 2 fixes:
 * don't rely on bytescale() if the data is uint8 => just do an add and mul (tiny bit slower?)
 * ensure irange is the same type as the data.