acq stream: Don't fail totally when drift correction causes ROI to go out of bounds
When acquiring an image with drift correction, it can happen that the drift causes the ROI
to go outside of the possible FoV.
Until now, the similar case with the anchor region was already handled, but not for the
ROI. It would then stop the whole acquisition with this unexplicit message:
IndexError: Trying to assign value '(2049.227786423397, 329.31877459603504)' outside of the range (-2048.0, -2048.0)->(2048.0, 2048.0).

Now we correctly detect it. An explicit error message is written, and the acqusition continues
with whatever points can still be acquired.