acq stream: disable data rescaling if the data is 8 bits
If data is 8 bits, it can be directly mapped to the RGB data.
So don't try to do auto B/C by default on such data.

Works both when it's live data, and when it's loaded data.

That's part of the work to connect the Hw auto B/C of the Phenom into the GUI.