gui cont stream: fix playing a fluo stream causing 2 streams to play
Because the fluostreams are so closely related to the optical state, a bug was
giving the possibility to activate 2 streams simultaneously:
1. select combined view
2. add second fluo stream
3. play fluo stream 1
4. play sem stream
5. play fluo stream 2
=> boom: fluo streams 1&2 are playing

That was because when turning on fluo stream 2, should_update of sem stream (going to pause)
cause the microscope state to be updated... which cause fluo stream 1 to be activated as
it was the last optical stream played... and finally fluo stream 2 is set to be activated.

=> don't call the scheduler multiple times when activating one stream, just disable
temporarily the scheduler subscriptions while it's running.