driver spec*: convert all the getPolyToWavelength() to getPixelToWavelength()
The easiest way to support calibration in SpectraPro is to just compute the wavelength for each pixel.
Note that it could still be represented by a polynomial (of second order) as Winspec does.
But anyway, at the end, we'd compute the values for each pixel.
So we just completely skip the polynomial part.

Then we can simplify quite a bit. This only does the driver side, as we still want to support polynomials for the
future, or to read old files.