script odemis-stop: really try not to kill backend that has just started
Drop pkill which was only useful when using sudoers, but now we gave up on that.

Compare the new and old PIDS and only kill and clean the running dir if some old PIDS are left.
That means that most of the time the run dir is not (forced) cleaned up.

That means that the new running backend might have a chance to actually work.