acq stream: reorganise Streams
Introduce LiveStream, super class of all the streams that can be updated live.
This avoids running a histogram thread in the StaticStreams.

Also separate spot mode from SEMStream => it's now a special stream SpotSEMStream.
This simplifies the code, and allows to specify the position of the spot.

Also force an update to the .image if the histogram has been updated.
This ensures that if auto_bc is on, the image is always with the right B/C.