gui cont view: ensure the stage position is displayed at init
On the SECOM, it's really strange to not see the position until we explicitly move.

On the Delphi, it was fine before as loading the stage would cause the stage to move.
So now we see a point even if the sample holder is not loaded. Hopefully it's not a big
deal. If it is, we could detect the sample holder is not present, and skip the position
update.