andorshrk: support having the CCD passed as child
Change the driver so that if it doesn't access the spectrometer via the I2C port
of the CCD, it still need one CCD as a child, to have getPixelToWavelength()
work properly.

Maybe in the future, if we have the spectrograph without CCD, we could also
support no child at all, but for now, it's either camera, of child = ccd.

Also add internal support for controlling the accessory port.