odemisd: terminate components in more clever order
First terminate components which do not depend on anything else (aka parents),
and latter the lower components in the hierarchy. That avoids some components
failing to terminate or failing while other terminate.

Still not perfect, as the children created by delagation are also terminated
by delegation and we do not take this into account.