acq stream: Do not update the local VA from the hw VAs
It's complicated, generally not useful for the user,
and in some cases where the stream (or something else) changes temporarily the hardware settings, that
messes up everything.
So it's better to have it just one way: local VA -> hw VA

That should fix some problems with the spot stream and monochromator stream on the SPARC.

Also rename onNewImage() to onNewData as this is called when new raw _data_ is received.

Also make sure that the local VAs are set before running the acquisition in the MDStream.
Should fix some major acquisition problem on the SPARC.