acq stream: various fixes with local VAs
Make sure that the MDStream use the local VAs (need to "link" them before acquisition)

Special hack for the .resolution local VAs which cause lots of troubles, especially when there is a ROI (cf SEMStream)

Fix setting order of the local VA (used wrong names)

Separate the acquisition time of the SettingsStream (= time for the live view) from the acquisition time of the MDStream (= time with repetition).

Small clean up of the assembleTile() method to look more like assembleMainData().

Fix some test cases.