gui cont streams: take charge of the spot stream
The stream controller is in charge of scheduling streams, it makes sense to also schedule
the spot stream. Especially, this simplifies a lot the code.

We now have just 2 kinds of streams: the one that need spot, and the one which cannot use spot.

=> Set the spot mode based on which type of stream is being played.

Still a little ugly in the scheduler => need to introduce a new policy? Or at least a method per policy.